Replace astropy with crates in make_regions_from_binmap
This does not replace all uses of astropy.io.fits, just those in the
make_regions_from_binmap routine. I fully expect this code to fail,
as the "trick" is in reading in the coordinate-transformation information
in the image files, and it can be tricky to support all the possible
ways of encoding this information in a FITS file (since crates takes a
"higher-level" approach than astropy.io.fits).